refactor: useSidemenu
