+++ src/lib/undux/utils/sidemenu.js
@@ -1,6 +1,5 @@
 // @flow
 import type { Store } from 'undux'
-import SimpleStore from '../SimpleStore'
 
 export const showSidemenu = (store: Store) => {
   store.set('sidemenu')({ visible: true })
@@ -12,9 +11,4 @@ export const hideSidemenu = (store: Store) => {
 
 export const toggleSidemenu = (store: Store) => {
   store.set('sidemenu')({ visible: !store.get('sidemenu').visible })
-}
-
-export const useSidemenu = () => {
-  const store = SimpleStore.useStore()
-  return [toggleSidemenu.bind(null, store), hideSidemenu.bind(null, store), showSidemenu.bind(null, store)]
 }
